+++ lib/flutter_dropdown_button.dart
@@ -4,6 +4,7 @@ export 'src/buttons/base_dropdown_button.dart';
 export 'src/buttons/basic_dropdown_button.dart';
 export 'src/buttons/dropdown_mixin.dart';
 export 'src/buttons/dynamic_text_base_dropdown_button.dart';
+export 'src/buttons/menu_alignment.dart';
 export 'src/buttons/text_only_dropdown_button.dart';
 export 'src/config/text_dropdown_config.dart';
 export 'src/dropdown_item.dart';